requirements.txt
pandas
requests
sqlalchemy
psycopg2
python-dotenv
spotipy